refactored profile card into different components
